requirements.txt
Flask>=2.0
flask-cors>=3.0.10
requests>=2.28
scikit-learn>=1.2
numpy>=1.23
pandas>=1.5
pytest>=7.0
python-dotenv>=1.0
geopy>=2.4